modified:   .travis.yml 	modified:   addons/base/views.py 	modified:   admin/base/settings/defaults.py 	modified:   admin/package.json 	modified:   admin/rdm_addons/oauth/views.py 	modified:   admin/rdm_addons/views.py 	modified:   admin/rdm_statistics/requirements.txt 	modified:   admin/rdm_statistics/views.py 	modified:   admin/rdm_timestampadd/views.py 	modified:   admin/static/js/rdm_addons/figshare/rdm-cfg.js 	modified:   admin/static/js/rdm_addons/owncloud/owncloudRdmConfig.js 	modified:   admin/static/js/rdm_addons/rdm-addons-page.js 	modified:   admin/static/js/rdm_addons/rdmAddonSettings.js 	modified:   admin/static/js/rdm_addons/s3/s3RdmConfig.js 	modified:   admin/templates/base.html 	modified:   admin/templates/rdm_addons/addon_list.html 	modified:   admin/yarn.lock 	modified:   admin_tests/rdm_addons/api_v1/test_views.py 	modified:   admin_tests/rdm_addons/oauth/test_views.py 	modified:   admin_tests/rdm_addons/test_views.py 	modified:   admin_tests/rdm_announcement/test_views.py 	modified:   admin_tests/rdm_statistics/test_views.py 	modified:   api/base/settings/local-travis.py 	modified:   api_tests/users/views/test_user_list.py 	modified:   osf/models/mixins.py 	new file:   root_cert_verifycate.pem 	modified:   tests/test_views.py 	modified:   website/oauth/views.py 	modified:   website/profile/views.py 	modified:   website/project/views/node.py 	modified:   website/project/views/timestamp.py

+++ admin/rdm_timestampadd/views.py
@@ -358,11 +358,11 @@ class AddTimestampData(RdmPermissionMixin, View):
                                                 download_file_path, tmp_dir)
             shutil.rmtree(tmp_dir)
         except Exception as err:
-            if os.path.exists(tmp_dir):
-                shutil.rmtree(tmp_dir)
-            raise ValueError('Exception:{}'.format(err))
-
-        request_data.update({'result': result})
+            if tmp_dir:
+                if os.path.exists(tmp_dir):
+                    shutil.rmtree(tmp_dir)
+        if 'result' in locals():
+            request_data.update({'result': result})
         return HttpResponse(json.dumps(request_data), content_type='application/json')
 
     def web_api_url(self, node_id):
